+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Expected: February 2024
 
 ### Corrected Bugs
 
+* Fixed: [Device connection errors could be more informative](https://github.com/clicon/clixon-controller/issues/7)
 * Fixed: [apply services does not allow dry-run](https://github.com/clicon/clixon-controller/issues/94)
 * Fixed: [old device config wrongly pushed back](https://github.com/clicon/clixon-controller/issues/93)
 * Fixed: [services reapply does not allow specific service instance](https://github.com/clicon/clixon-controller/issues/80)